refactor: Reorganized header components
