+++ src/components/Header/AppHeader.tsx
@@ -1,30 +1,10 @@
-import React, { ReactElement } from "react";
 import styled from "styled-components";
 
-import { AicsLogoSVG } from "../../assets";
-import { FlexRowAlignCenter } from "../../styles/utils";
-
-const AicsLogoLink = styled.a`
-  position: relative;
-  width: 180px;
-  height: 46px;
-`;
-
-const StyledAicsLogo = styled(AicsLogoSVG)`
-  left: 0;
-  top: 0;
-  position: absolute;
-`;
-
-const VerticalDivider = styled.div`
-  height: 24px;
-  width: 1px;
-  background-color: var(--color-dividers);
-  display: inline-block;
-`;
-
-/** Top title bar for the app */
-export const AppHeader = styled.div`
+/**
+ * Top title bar for the app, which will stick to the top of the page.
+ * Child components will be spaced apart evenly.
+ * */
+const Header = styled.div`
   display: flex;
   flex-direction: row;
   flex-wrap: wrap;
@@ -43,14 +23,4 @@ export const AppHeader = styled.div`
   left: 0;
 `;
 
-export function HeaderLogo(): ReactElement {
-  return (
-    <FlexRowAlignCenter $gap={20}>
-      <AicsLogoLink href="https://www.allencell.org/" rel="noopener noreferrer">
-        <StyledAicsLogo title={"https://www.allencell.org"} />
-      </AicsLogoLink>
-      <VerticalDivider />
-      <h1 style={{ whiteSpace: "nowrap" }}>Timelapse Colorizer</h1>{" "}
-    </FlexRowAlignCenter>
-  );
-}
+export default Header;
