+++ src/components/Header/index.tsx
@@ -1,0 +1,4 @@
+import Header from "./AppHeader";
+import HeaderLogo from "./HeaderLogo";
+
+export { HeaderLogo, Header };
